changes added for settings page
